+++ node/clients/index.ts
@@ -2,6 +2,8 @@ import { IOClients } from '@vtex/api'
 
 import LogisticsClient from './logistics'
 import Status from './status'
+import CustomDataDataManager from './customDataManager'
+import FileManager from './fileManager'
 
 // Extend the default IOClients implementation with our own custom clients.
 export class Clients extends IOClients {
@@ -11,5 +13,13 @@ export class Clients extends IOClients {
 
   public get logistics() {
     return this.getOrSet('logistics', LogisticsClient)
+  }
+
+  public get customDataManager() {
+    return this.getOrSet('customDataManager', CustomDataDataManager)
+  }
+
+  public get fileManager() {
+    return this.getOrSet('fileManager', FileManager)
   }
 }