+++ node/constants/index.ts
@@ -1,5 +1,6 @@
 export const BUCKET = 'imageprotocol'
-export const CONFIG_PATH = 'mappings'
 export const CONFIG_PATH_CCPOLYGON = 'customerclass-polygon'
 export const CONFIG_PATH_CC = 'customerclass'
 export const CONFIG_PATH_POLYGON = 'polygon'
+export const POLYGON_PRIORITY_SETTING = 'polygonPriority'
+export const CUSTOMER_CLASS_PRIORITY_SETTING = 'customerClassPriority'