+++ node/index.ts
@@ -1,21 +1,26 @@
-// import type { ClientsConfig } from '@vtex/api'
-import type { ClientsConfig, ServiceContext } from '@vtex/api'
-import { method, Service } from '@vtex/api'
+import type { ClientsConfig, RecorderState, ServiceContext } from '@vtex/api'
+import { LRUCache, method, Service } from '@vtex/api'
 
-import { errorHandler, getImgUrl } from './middlewares'
+import {
+  errorHandler,
+  getImgUrl,
+  getUsersPolygon,
+  getUserCustomerClass,
+  validations,
+  saveInfo,
+  deleteRecord,
+} from './middlewares'
 import { saveDataInfo } from './resolvers/saveDataInfo'
 import { getPolygons } from './resolvers/getPolygons'
 import { getDataList } from './resolvers/getDataList'
 import { removeFromList } from './resolvers/removeFromList'
 import { Clients } from './clients'
-// import { status } from './middlewares/status'
-// import { validate } from './middlewares/validate'
 
 // Create a LRU memory cache for the Status client.
 // The @vtex/api HttpClient respects Cache-Control headers and uses the provided cache.
-// const memoryCache = new LRUCache<string, any>({ max: 5000 })
+const memoryCache = new LRUCache<string, any>({ max: 5000 })
 
-// metrics.trackCache('status', memoryCache)
+metrics.trackCache('status', memoryCache)
 
 // This is the configuration for clients available in `ctx.clients`.
 const TIMEOUT_MS = 2000
@@ -27,25 +32,37 @@ const clients: ClientsConfig<Clients> = {
       retries: 0,
       timeout: TIMEOUT_MS,
     },
+    status: {
+      memoryCache,
+    },
   },
 }
 
 declare global {
+  interface DataRequest {
+    fields: Fields
+    files: any[] | undefined
+  }
+  interface Fields {
+    [key: string]: any
+  }
+  interface State extends RecorderState {
+    customerClass: string | undefined
+    polygons: string[] | undefined
+    request: DataRequest
+  }
+
   // We declare a global Context type just to avoid re-writing ServiceContext<Clients, State> in every handler and resolver
-  type Context = ServiceContext<Clients>
+  type Context = ServiceContext<Clients, State>
 
   interface ClientMasterdataEntityType {
     customerClass: string
     where: string
   }
-  // The shape of our State object found in `ctx.state`. This is used as state bag to communicate between middlewares.
-  // interface State extends RecorderState {
-  //   code: number
-  // }
 }
 
 // Export a service that defines route handlers and client options.
-export default new Service({
+export default new Service<Clients, State, Context>({
   clients,
   graphql: {
     resolvers: {
@@ -60,6 +77,10 @@ export default new Service({
     },
   },
   routes: {
-    getUrl: method({ GET: [errorHandler, getImgUrl] }),
+    getUrl: method({
+      GET: [errorHandler, getUsersPolygon, getUserCustomerClass, getImgUrl],
+    }),
+    saveInfo: method({ POST: [validations, saveInfo] }),
+    deleteRecord: method({ DELETE: [deleteRecord] }),
   },
 })