+++ node/typings/requestData.d.ts
@@ -1,0 +1,11 @@
+interface RequestData {
+  userId: string
+  imageProtocolId: string
+  latitude: string
+  longitude: string
+}
+
+interface CustomerData {
+  customerClass?: string | undefined
+  polygons?: string[] | undefined
+}